+++ NEWS.md
@@ -6,6 +6,9 @@
 ### Bug fixes
 - Reddit spout allows wider range of URLs, including absolute URLs and searches ([#1033](https://github.com/SSilence/selfoss/pull/1033))
 
+### Other changes
+- Removed broken instapaper scraping from Reddit spout ([#1033](https://github.com/SSilence/selfoss/pull/1033))
+
 
 ## 2.18 – 2018-03-05
 ### New features